linebot-Emotion/cloudrun-webhook/requirements.txt
gunicorn
flask>=2.3.0
flask-cors>=1.10.0         # ✅ 加上這行
gunicorn>=21.0.0
python-dotenv>=1.0.0
google-cloud-language>=2.12.0
google-generativeai>=0.2.0 # ✅ 加上這行
firebase-admin>=6.0.0
matplotlib>=3.7.0
pandas>=1.5.0
numpy>=1.23.0
requests>=2.25.0
